+++ src/aprogression.rs
@@ -1,12 +1,12 @@
 use crate::common::sigma;
 
-pub fn sum(a1: f32, d: f32, n: usize) -> f32 {
+pub fn sum(a1: f64, d: f64, n: usize) -> f64 {
     //! sum = a1 * n + d * sigma(n - 1)
 
-    a1 * n as f32 + d * sigma(n - 1) as f32
+    a1 * n as f64 + d * sigma(n - 1) as f64
 }
 
-pub fn get_n(a1: f32, d: f32, sum: f32) -> usize {
+pub fn get_n(a1: f64, d: f64, sum: f64) -> usize {
     //! n - number of iterations will be spent before 'sum' is achieved
 
     let mut n = Box::<usize>::new(1);
@@ -16,11 +16,11 @@ pub fn get_n(a1: f32, d: f32, sum: f32) -> usize {
     *n
 }
 
-pub fn get_a_n(a_k: f32, d: f32, k: usize, n: usize) -> f32 {
+pub fn get_a_n(a_k: f64, d: f64, k: usize, n: usize) -> f64 {
     //! a_n = a_k + (n - k) * d , n => k
     //!
     //! a_n = a_k - (k - n) * d , n <= k
 
-    if n > k { return a_k + (n - k) as f32 * d }
-    a_k - (k - n) as f32 * d
+    if n > k { return a_k + (n - k) as f64 * d }
+    a_k - (k - n) as f64 * d
 }
