+++ src/vector_algebra.rs
@@ -2,12 +2,12 @@ use std::error::Error;
 
 #[derive(PartialEq, Debug)]
 pub struct Vector3D {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 #[allow(dead_code)]
-pub fn cos_alpha(a: &Vector3D, b: &Vector3D) -> Result<f32, Box<dyn Error>> {
+pub fn cos_alpha(a: &Vector3D, b: &Vector3D) -> Result<f64, Box<dyn Error>> {
     let ab = a.x * b.x + a.y * b.y + a.z * b.z;
     let a = (a.x.powf(2.0) + a.y.powf(2.0) + a.z.powf(2.0)).sqrt();
     let b = (b.x.powf(2.0) + b.y.powf(2.0) + b.z.powf(2.0)).sqrt();
