+++ tests/test_func_analysis.rs
@@ -6,7 +6,7 @@ mod test_func_analysis {
     #[test]
     fn test_find_roots1_pos() {
         let mut y = Expression::new(
-            |x: f32| x.powf(3.0) - 16.0 * x.powf(2.0) / 3.0 + 15.0 * x
+            |x: f64| x.powf(3.0) - 16.0 * x.powf(2.0) / 3.0 + 15.0 * x
         );
         y.find_roots();
         let root = y.roots()[0].unwrap();
@@ -16,7 +16,7 @@ mod test_func_analysis {
     #[test]
     fn test_find_roots2_none_pos() {
         let mut y = Expression::new(
-            |x: f32| (x.powf(3.0) + 1.0) * (x - 1.0) / (x - 2.0).sqrt()
+            |x: f64| (x.powf(3.0) + 1.0) * (x - 1.0) / (x - 2.0).sqrt()
         );
         y.find_roots();
         let root = y.roots()[0];
@@ -26,7 +26,7 @@ mod test_func_analysis {
     #[test]
     fn test_find_extremums1_max_min_pos() {
         let mut y = Expression::new(
-            |x: f32| 6.0 * x.powf(5.0) - 90.0 * x.powf(3.0) - 5.0
+            |x: f64| 6.0 * x.powf(5.0) - 90.0 * x.powf(3.0) - 5.0
         );
         y.find_extremums(-5.0, 1.0);
 
@@ -44,7 +44,7 @@ mod test_func_analysis {
     #[test]
     fn test_find_extremums2_extremums_pos() {
         let mut y = Expression::new(
-            |x: f32| x.powf(3.0) - 16.0 * x.powf(2.0) / 3.0 + 15.0 * x
+            |x: f64| x.powf(3.0) - 16.0 * x.powf(2.0) / 3.0 + 15.0 * x
         );
         let res = y.find_extremums(-5.0, 1.0);
         let mut test_data = HashMap::<String, Point>::new();
@@ -57,7 +57,7 @@ mod test_func_analysis {
     #[test]
     fn test_find_extremums3_extremums_pos() {
         let mut y = Expression::new(
-            |x: f32| x.powf(2.0) + x
+            |x: f64| x.powf(2.0) + x
         );
         let res = y.find_extremums(-5.0, 1.0);
         let mut test_data = HashMap::<String, Point>::new();
@@ -70,7 +70,7 @@ mod test_func_analysis {
     #[test]
     fn test_find_extremums4_none_pos() {
         let mut y = Expression::new(
-            |x: f32| (x.powf(3.0) + 1.0) * (x - 1.0) / (x - 2.0).sqrt()
+            |x: f64| (x.powf(3.0) + 1.0) * (x - 1.0) / (x - 2.0).sqrt()
         );
         let res = y.find_extremums(-5.0, 1.0);
         assert!(res.is_none());
@@ -80,7 +80,7 @@ mod test_func_analysis {
     #[test]
     fn test_find_extremums5_inverted_args_pos() {
         let mut y = Expression::new(
-            |x: f32| (x.powf(3.0) + 1.0) * (x - 1.0) / (x - 2.0).sqrt()
+            |x: f64| (x.powf(3.0) + 1.0) * (x - 1.0) / (x - 2.0).sqrt()
         );
         let res = y.find_extremums(1.0, -5.0);
         assert!(res.is_none());
